+++ packages/taiko-client-rs/crates/protocol/src/shasta/mod.rs
@@ -5,6 +5,7 @@ pub mod constants;
 pub mod error;
 pub mod manifest;
 pub mod payload_helpers;
+pub mod rpc_methods;
 
 pub use blob_coder::BlobCoder;
 pub use error::{ForkConfigResult, ProtocolError, Result, ShastaForkConfigError};
@@ -12,3 +13,4 @@ pub use payload_helpers::{
     calculate_shasta_difficulty, compute_build_payload_args_id, encode_extra_data,
     encode_transactions, encode_tx_list,
 };
+pub use rpc_methods::DriverRpcMethod;
